Made a component
Created a component for the nav bar, so it looks a lot neater
